+++ providers/denominator-designate/src/test/java/denominator/designate/DesignateRecordSetLiveTest.java
@@ -1,0 +1,16 @@
+package denominator.designate;
+
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+import denominator.BaseRecordSetLiveTest;
+
+@Test
+public class DesignateRecordSetLiveTest extends BaseRecordSetLiveTest {
+    @BeforeClass
+    private void setUp() {
+        DesignateConnection connection = new DesignateConnection();
+        manager = connection.manager;
+        setMutableZoneIfPresent(connection.mutableZone);
+    }
+}